requirements.txt
# =========================
# Core API
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# =========================
# Config / Validation
# =========================
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
pydantic[email]>=2.7.4,<3

# =========================
# Database
# =========================
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# =========================
# Authentication
# =========================
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# =========================
# Background Jobs
# =========================
celery==5.3.6
redis==5.0.1

# =========================
# Storage
# =========================
boto3==1.34.34
google-cloud-storage==2.18.2

# =========================
# LangChain Ecosystem
# =========================
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-openai==0.2.6
langgraph==0.2.39

# =========================
# Vector Database
# =========================
chromadb==0.4.24

# =========================
# LLM Provider
# =========================
openai==1.55.3

# =========================
# HTTP / Retry
# =========================
httpx==0.26.0
tenacity==8.2.3

# =========================
# Logging
# =========================
structlog==24.1.0

# =========================
# File Processing
# =========================
pymupdf==1.23.21
pandas==2.2.0
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
pillow==10.4.0
pytesseract==0.3.10
extract-msg==0.55.0
beautifulsoup4==4.12.3
lxml==5.3.0

# =========================
# Testing
# =========================
pytest==8.0.0
pytest-asyncio==0.23.5
